\name{rOptManifold-package}
\alias{rOptManifold-package}
\alias{rOptManifold}
\docType{package}
\title{
\packageTitle{rOptManifold}
}
\description{
\packageDescription{rOptManifold}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{rOptManifold}
\packageIndices{rOptManifold}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{rOptManifold}

Maintainer: \packageMaintainer{rOptManifold}
}
\references{
}
\keyword{ package }
\seealso{
}
\examples{
}
